Keep the eviction decision out of the file that imports Playwright
CI cannot import `profiles.ts`: it pulls Playwright in at module scope, and the browsers are not
installed for that package there. This is the reason `authorisation.ts` and `bot-id.ts` already sit in
files of their own, and the eviction policy belongs in one for the same reason.

Caught by the `tests` job rather than locally, where Playwright happens to be installed.
